+++ src/app/src/util/constants.js
@@ -1,5 +1,7 @@
 export const OTHER = 'Other';
 
+export const FACILITIES_REQUEST_PAGE_SIZE = 25;
+
 // This choices must be kept in sync with the identical list
 // kept in the Django API's models.py file
 export const contributorTypeOptions = Object.freeze([